TST: Merged topography1 and topography2 tests
